# crestdb-data/src/gen/java/hep/crest/swagger/model/IovPayloadDto.java
/*
 * CrestDB REST API
 * OpenApi3 for conditions data management
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package hep.crest.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.math.BigDecimal;
import java.time.OffsetDateTime;
import com.fasterxml.jackson.annotation.JsonPropertyOrder;
import javax.validation.constraints.*;
import javax.validation.Valid;

/**
 * IovPayloadDto
 */
@JsonPropertyOrder({
  IovPayloadDto.JSON_PROPERTY_SINCE,
  IovPayloadDto.JSON_PROPERTY_INSERTION_TIME,
  IovPayloadDto.JSON_PROPERTY_VERSION,
  IovPayloadDto.JSON_PROPERTY_OBJECT_TYPE,
  IovPayloadDto.JSON_PROPERTY_SIZE,
  IovPayloadDto.JSON_PROPERTY_PAYLOAD_HASH,
  IovPayloadDto.JSON_PROPERTY_STREAMER_INFO
})
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaJerseyServerCodegen")
public class IovPayloadDto   {
  public static final String JSON_PROPERTY_SINCE = "since";
  @JsonProperty(JSON_PROPERTY_SINCE)
  private BigDecimal since;

  public static final String JSON_PROPERTY_INSERTION_TIME = "insertionTime";
  @JsonProperty(JSON_PROPERTY_INSERTION_TIME)
  private OffsetDateTime insertionTime;

  public static final String JSON_PROPERTY_VERSION = "version";
  @JsonProperty(JSON_PROPERTY_VERSION)
  private String version;

  public static final String JSON_PROPERTY_OBJECT_TYPE = "objectType";
  @JsonProperty(JSON_PROPERTY_OBJECT_TYPE)
  private String objectType;

  public static final String JSON_PROPERTY_SIZE = "size";
  @JsonProperty(JSON_PROPERTY_SIZE)
  private Integer size;

  public static final String JSON_PROPERTY_PAYLOAD_HASH = "payloadHash";
  @JsonProperty(JSON_PROPERTY_PAYLOAD_HASH)
  private String payloadHash;

  public static final String JSON_PROPERTY_STREAMER_INFO = "streamerInfo";
  @JsonProperty(JSON_PROPERTY_STREAMER_INFO)
  private String streamerInfo;

  public IovPayloadDto since(BigDecimal since) {
    this.since = since;
    return this;
  }

  /**
   * Get since
   * @return since
   **/
  @JsonProperty("since")
  @ApiModelProperty(value = "")
  @Valid 
  public BigDecimal getSince() {
    return since;
  }

  public void setSince(BigDecimal since) {
    this.since = since;
  }

  public IovPayloadDto insertionTime(OffsetDateTime insertionTime) {
    this.insertionTime = insertionTime;
    return this;
  }

  /**
   * Get insertionTime
   * @return insertionTime
   **/
  @JsonProperty("insertionTime")
  @ApiModelProperty(value = "")
  
  public OffsetDateTime getInsertionTime() {
    return insertionTime;
  }

  public void setInsertionTime(OffsetDateTime insertionTime) {
    this.insertionTime = insertionTime;
  }

  public IovPayloadDto version(String version) {
    this.version = version;
    return this;
  }

  /**
   * Get version
   * @return version
   **/
  @JsonProperty("version")
  @ApiModelProperty(value = "")
  
  public String getVersion() {
    return version;
  }

  public void setVersion(String version) {
    this.version = version;
  }

  public IovPayloadDto objectType(String objectType) {
    this.objectType = objectType;
    return this;
  }

  /**
   * Get objectType
   * @return objectType
   **/
  @JsonProperty("objectType")
  @ApiModelProperty(value = "")
  
  public String getObjectType() {
    return objectType;
  }

  public void setObjectType(String objectType) {
    this.objectType = objectType;
  }

  public IovPayloadDto size(Integer size) {
    this.size = size;
    return this;
  }

  /**
   * Get size
   * @return size
   **/
  @JsonProperty("size")
  @ApiModelProperty(value = "")
  
  public Integer getSize() {
    return size;
  }

  public void setSize(Integer size) {
    this.size = size;
  }

  public IovPayloadDto payloadHash(String payloadHash) {
    this.payloadHash = payloadHash;
    return this;
  }

  /**
   * Get payloadHash
   * @return payloadHash
   **/
  @JsonProperty("payloadHash")
  @ApiModelProperty(value = "")
  
  public String getPayloadHash() {
    return payloadHash;
  }

  public void setPayloadHash(String payloadHash) {
    this.payloadHash = payloadHash;
  }

  public IovPayloadDto streamerInfo(String streamerInfo) {
    this.streamerInfo = streamerInfo;
    return this;
  }

  /**
   * Get streamerInfo
   * @return streamerInfo
   **/
  @JsonProperty("streamerInfo")
  @ApiModelProperty(value = "")
  
  public String getStreamerInfo() {
    return streamerInfo;
  }

  public void setStreamerInfo(String streamerInfo) {
    this.streamerInfo = streamerInfo;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    IovPayloadDto iovPayloadDto = (IovPayloadDto) o;
    return Objects.equals(this.since, iovPayloadDto.since) &&
        Objects.equals(this.insertionTime, iovPayloadDto.insertionTime) &&
        Objects.equals(this.version, iovPayloadDto.version) &&
        Objects.equals(this.objectType, iovPayloadDto.objectType) &&
        Objects.equals(this.size, iovPayloadDto.size) &&
        Objects.equals(this.payloadHash, iovPayloadDto.payloadHash) &&
        Objects.equals(this.streamerInfo, iovPayloadDto.streamerInfo);
  }

  @Override
  public int hashCode() {
    return Objects.hash(since, insertionTime, version, objectType, size, payloadHash, streamerInfo);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class IovPayloadDto {\n");
    
    sb.append("    since: ").append(toIndentedString(since)).append("\n");
    sb.append("    insertionTime: ").append(toIndentedString(insertionTime)).append("\n");
    sb.append("    version: ").append(toIndentedString(version)).append("\n");
    sb.append("    objectType: ").append(toIndentedString(objectType)).append("\n");
    sb.append("    size: ").append(toIndentedString(size)).append("\n");
    sb.append("    payloadHash: ").append(toIndentedString(payloadHash)).append("\n");
    sb.append("    streamerInfo: ").append(toIndentedString(streamerInfo)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}